feat: Completely refactor the `FrameProcessorPlugins.ts` file into multiple files (#2830)
* feat: Completely refactor the `FrameProcessorPlugins.ts` file into multiple files

* Refactor

* Prepare

* Update some docs

* fix: Fix invalid `options` param (undefined)

* Update FRAME_PROCESSORS_CREATE_OVERVIEW.mdx
