+++ docs/docs/guides/FRAME_PROCESSORS_CREATE_OVERVIEW.mdx
@@ -64,7 +64,7 @@ Returns a `string` in JS:
 ```js
 export function detectObject(frame: Frame): string {
   'worklet'
-  const result = FrameProcessorPlugins.detectObject(frame)
+  const result = detectObject(frame)
   console.log(result) // <-- "cat"
 }
 ```
