+++ docs/docs/guides/FRAME_PROCESSOR_CREATE_FINAL.mdx
@@ -9,9 +9,9 @@ sidebar_label: Finish creating your Frame Processor Plugin
 To make the Frame Processor Plugin available to the Frame Processor Worklet Runtime, create the following wrapper function in JS/TS:
 
 ```ts
-import { VisionCameraProxy, Frame } from 'react-native-vision-camera'
+import { initFrameProcessorPlugin, Frame } from 'react-native-vision-camera'
 
-const plugin = VisionCameraProxy.initFrameProcessorPlugin('scanFaces')
+const plugin = initFrameProcessorPlugin('scanFaces')
 
 /**
  * Scans faces.
