+++ package/src/Camera.tsx
@@ -8,7 +8,7 @@ import { CameraModule } from './NativeCameraModule'
 import type { PhotoFile, TakePhotoOptions } from './types/PhotoFile'
 import type { Point } from './types/Point'
 import type { RecordVideoOptions, VideoFile } from './types/VideoFile'
-import { VisionCameraProxy } from './FrameProcessorPlugins'
+import { VisionCameraProxy } from './frame-processors/VisionCameraProxy'
 import { CameraDevices } from './CameraDevices'
 import type { EmitterSubscription, NativeSyntheticEvent, NativeMethods } from 'react-native'
 import type { TakeSnapshotOptions } from './types/Snapshot'
