+++ package/src/index.ts
@@ -1,7 +1,8 @@
+// Base Camera Exports
 export * from './Camera'
 export * from './CameraError'
-export * from './FrameProcessorPlugins'
 
+// Types
 export * from './types/CameraDevice'
 export * from './types/CameraProps'
 export * from './types/Frame'
@@ -13,10 +14,12 @@ export * from './types/Point'
 export * from './types/VideoFile'
 export * from './types/CodeScanner'
 
+// Devices API
 export * from './devices/getCameraFormat'
 export * from './devices/getCameraDevice'
 export * from './devices/Templates'
 
+// Hooks
 export * from './hooks/useCameraDevice'
 export * from './hooks/useCameraDevices'
 export * from './hooks/useCameraFormat'
@@ -24,4 +27,12 @@ export * from './hooks/useCameraPermission'
 export * from './hooks/useCodeScanner'
 export * from './hooks/useFrameProcessor'
 
+// Frame Processors
+export * from './frame-processors/initFrameProcessorPlugin'
+export * from './frame-processors/runAsync'
+export * from './frame-processors/runAtTargetFps'
+// DEPRECATED: This will be removed in favour of a CxxTurboModule in the future.
+export * from './frame-processors/VisionCameraProxy'
+
+// Skia Frame Processors
 export * from './skia/useSkiaFrameProcessor'
